+++ resources/ReleaseNotes.md
@@ -1,11 +1,13 @@
 # Software
 
 - Export I-ALiRT data
+- Separate `ScienceProcessing` and `IALiRTProcessing` in `mag.IMAPTestingAnalysis`
 - Add operations to plot composite variables (e.g., difference between variables)
 - Add plot for comparison of I-ALiRT data to science data
 - Add optional global legend to figures generated with `mag.graphics.visualize`
 - Add option to change tile indexing in figures generated with `mag.graphics.visualize`
 - Also crop I-ALiRT data when calling `cropScience`
 - Remove check of mistimed packets
+- Remove logic to handle old versions of `mag.IMAPTestingAnalysis` in MAT files
 - Fix wrong values for I-ALiRT data and packet frequency
 - Fix issue with too many sensor events being filtered out when processing science